divide-and-conquer!
